Fix Revenue number display, add M1 columns to table, persist column visibility, clickable dashboard jobs
- Backend: Change Decimal to float in RevenueRow/RevenueStatement models so JSON
  serializes as numbers instead of strings (was breaking all frontend arithmetic)
- Revenue: Replace simplified table columns with full M1 export columns (29 cols)
  with sensible defaults visible; add Number() coercion for old Firestore data
- All tools: Persist column visibility in localStorage so preferences survive
  page navigation and browser refresh
- Proration: Add hidden columns (interest type, legal desc, property, operator,
  block, section, abstract, well type, notes) toggleable via column picker
- Dashboard: Make recent activity rows clickable (navigates to tool page),
  add status column showing completed/failed with row counts

+++ backend/app/models/revenue.py
@@ -1,7 +1,6 @@
 """Pydantic models for Revenue PDF extraction tool."""
 
 from datetime import date
-from decimal import Decimal
 from enum import Enum
 from typing import Optional
 
@@ -24,18 +23,18 @@ class RevenueRow(BaseModel):
     sales_date: Optional[date] = None
     product_code: Optional[str] = None
     product_description: Optional[str] = None
-    decimal_interest: Optional[Decimal] = None
+    decimal_interest: Optional[float] = None
     interest_type: Optional[str] = None
-    avg_price: Optional[Decimal] = None
-    property_gross_volume: Optional[Decimal] = None
-    property_gross_revenue: Optional[Decimal] = None
-    owner_volume: Optional[Decimal] = None
-    owner_value: Optional[Decimal] = None
-    owner_tax_amount: Optional[Decimal] = None
+    avg_price: Optional[float] = None
+    property_gross_volume: Optional[float] = None
+    property_gross_revenue: Optional[float] = None
+    owner_volume: Optional[float] = None
+    owner_value: Optional[float] = None
+    owner_tax_amount: Optional[float] = None
     tax_type: Optional[str] = None
-    owner_deduct_amount: Optional[Decimal] = None
+    owner_deduct_amount: Optional[float] = None
     deduct_code: Optional[str] = None
-    owner_net_revenue: Optional[Decimal] = None
+    owner_net_revenue: Optional[float] = None
 
 
 class RevenueStatement(BaseModel):
@@ -45,7 +44,7 @@ class RevenueStatement(BaseModel):
     format: StatementFormat
     payor: Optional[str] = None
     check_number: Optional[str] = None
-    check_amount: Optional[Decimal] = None
+    check_amount: Optional[float] = None
     check_date: Optional[date] = None
     operator_name: Optional[str] = None
     owner_number: Optional[str] = None
